parser: move parser and scanner in vcsn::rat_exp.
* vcsn/io/parse-rat-exp.yy (namespace, name-prefix): Define
appropriately so that parser and scanner be in vcsn::rat_exp.
(YYSTYPE): Do not define by hand, rather use...
(%union): this, so that we can move it to a better place in
the input file, and so that it is in the right place in the
output file.
Remove now unneeded namespace qualifications.
Adjust %code directives.
Remove now useless typedef to shorten access into vcsn::rat_exp.
* vcsn/io/lexer.ll (TOK): Adjust.