rat: be parameterized by Weight, not WeightSet.
After discussion with Alexandre Duret-Lutz, it is decided that
the rat exp bare values, but not their interpretation.  The
WeightSet aspect (including rules of computations over the
weights) is therefore transfered to the factory, not the ast.

* vcsn/core/rat/node.fwd.hh, vcsn/core/rat/node.hh,
* vcsn/core/rat/node.hxx: Parameterize per Weight, not WeightSet.
Adjust all the typedefs.
As a consequence, the nodes no longer know the unit of the WeightSet,
so they must be explicitly given their weights.
Change all the constructors accordingly.

* vcsn/core/rat/factory.hh, vcsn/core/rat/factory.hxx: Adjust to
properly pass the needed weights.

* vcsn/core/rat/printer.hh, vcsn/core/rat/printer.hxx,
vcsn/core/rat/visitor.hh, vcsn/core/rat/visitor.hxx,
vcsn/io/parse-rat-exp.yy: Adjust the parameters.

There is one dead body: we now always display the weights,
including if it is the unit.  This is because currently the
printer does not know the WeightSet.  This will be addressed
later.