rat: factory: split the abstract factory files from the concrete ones.
There are too many mutual dependencies, especially if
we want to be able to have RatExps with weights which
are RatExps, since in that case the parser needs the
factory and the factory needs the parser (to convert
weights as strings into RatExp).

Besides, this move makes it easier to have a compiled
parser that it not parameterized at all: just pass it
a concrete factory.

* vcsn/core/rat/factory.hh, vcsn/core/rat/factory.hxx: Extract
concrete factory implementation into...
* vcsn/core/rat/factory_.hh, vcsn/core/rat/factory_.hxx: here.