rat: parser: revamp the handling of the weights.
The current interface is a real mess.  There are way too many
functions called "weight", and weirdly enough, instead of having
simple functions on lists of weights that bounce to functions on
weights, almost all the computations are performed in the former.

Note also that this set-up, which delays the parsing of the weights
very late, does not provide an easy means to provide accurate location
for syntax errors in weights.

Besides, there are many identities that are not properly implemented.

In order to fix all this, the following commit changes the handling of
the weights in the parser: instead of storing then as a list of
strings, store them as the current weight of an identity.  So weights
are converted immediately, there products are performed immediately
too (no need for a list).  Then, associating the weight is just a
regular product.

* vcsn/core/rat/factory.hh, vcsn/core/rat/factory.hxx
(weight_str_container, weight_str, weight): Remove.
* vcsn/core/rat/factory_.hh, vcsn/core/rat/factory_.hxx (weight):
Remove the old ones, keep only four overloads: left/right
scalar product on abstract/concrete type.
* vcsn/io/parse-rat-exp.yy: Adjust.
* vcsn/misc/cast.hh (maybe_down_cast): New.