pprat: expose static types.
In a forthcoming patch, initial support for "standard_of" is started.
But it shows that static types must be exposed, currently way too much
is lost with "factory*" (as opposed to "factory_<...>".  So to prepare
this, expose the different factories _with their types_.

* tests/rat/pprat.cc (fact_b, fact_br, fact_z, fact_zr, fact_zrr):
New.
(pprat): Be templated.
(pprat): New overload, to bounce on the templated one.
(main): Adjust.
(alpha_t, alpha): Be global (sorry...).