kratexp: make the expression constants.
We've had too many problems because we modified the expressions.
Make them const.

Refactoring is called for... For instance in the area of the casts.
Not to mention that static_pointer_cast should suffice, instead of
dynamic_pointer_cast.

* vcsn/core/rat/node.fwd.hh (exp_t): Be const.
* vcsn/core/rat/node.hh: Populate with "const".
(self_t, shared_t, clone): New.
There are many opportunities for refactoring :(
Especially if macros are accepted ;)
* vcsn/core/rat/kratexps.hxx: Adjust.
Handle more trivial identities in the exp level (independant of the
weight set).