c++: beware of lambdas
One of our function used to return a lambda which captured by
reference an argument of the enclosing function.  As a result, the
lambda was using a reference to a dead value.

No longer use closure with references, use copies only.  And actually,
explicitly list all the captured arguments, instead of relying on [=]
and the like.

* vcsn/core/entryiter.hh, vcsn/core/mutable_automaton.hh: here
Avoid useless rvalue-ref.