kratexp: really accept only the right label_t.
We used to always take "word_t" as incoming type to build atoms.  It
is inside these routines that it was checked that in the case of
labels_are_letters, then the word was actually a single letter.

Move this to the abstract_kratexpset level, so that the kratexpset
level is clearly taking only the right type.

* vcsn/core/rat/abstract_kratexpset.hh: Handle the conversion word_t
to letter_t if needed.
* vcsn/core/rat/kratexpset.hh, vcsn/core/rat/kratexpset.hxx (atom): Now
accept only the right label_t.
* vcsn/algos/lift.hh: Simplify accordingly.