lae: do not store the labels in the automata
The tuple to store transition is already designed to not store labels
when they are "void".  This code was not used (as proved by the
multiple calls to "tr.label", which is not defined in that case).

Instead of using "void", I chose to use an empty type, in particular
so that the signature remain equal with or without stored labels: just
pass {} to mean "the empty label", whereas "void" cannot be used for
arguments.

* vcsn/empty.hh: New.
* vcsn/alphabets/char.hh, vcsn/Makefile.am: Adjust.
* vcsn/core/transition.hh: Instead of special-casing the "void" label_t,
handle empty_t.
Add accessors for labels.
* vcsn/core/mutable_automaton.hh: Adjust.