mutable_automaton::add_entry and del_entry
Discussed with Alexandre Duret-Lutz.  It is unclear what set_entry
should do (should certainly be equivalent to del_entry + add_entry),
and it will probably not be fast, so leave it aside for the time
being.

I really do not understand why entry_t is forcing the use of word_t as
a label.  It it more troublesome than helping in my humble opinion.

* vcsn/core/mutable_automaton.hh (add_entry, add_entry_): New.
Looks pretty useless, but is actually helpful as it will allow to
use abstract_polynomial to denote entries in an abstract way
(i.e., without Context parameter).
This is most useful...
* lib/vcsn/dot/parse.yy: here, so that we can, progressively,
get rid of exact types, and move to abstract types.

* vcsn/core/mutable_automaton (del_entry): New, for consistency.
* tests/unit/aut_char_z_lal.cc: Exercise them.

* NEWS.txt: Update.