introduce abstract_context
This is most useful to be able to manipulate an abstract automaton:
we want to be able to know its kind (sort of our own typeid).

We could do with a virtual method in the automata, yet I feel it will
help us to have the feature available directly from the context
itself.

* vcsn/ctx/ctx.hh (abstract_context): New.
(name): Now must be a regular function member, no longer a static one.
Adjust all dependencies.
(genset_string): New.
* vcsn/core/rat/kratexpset.hh, vcsn/ctx/fwd.hh, vcsn/weights/poly.hh:
Adjust.
* vcsn/algos/dotty.hh: Use context::name and genset_string instead
of ad-hoc code.