vname and sname.
First we had static members to get the "type name", which sounded
correct, since they will be used as keys to store algorithms for
_types_, and types are by essence static.

Yet we also need to have dynamic type names when facing an abstract
object (e.g., an abstract_mutable_automaton) for which we are looking
for a given algorithm.

That prompted me to convert all the static type name members to be
dynamic.  I now realize we need both: static names are useful for
instance to register algorithms, and dynamic names to query
algorithms.

* lib/vcsn/algos/edit-automaton.cc, lib/vcsn/algos/standard-of.cc,
* vcsn/algos/dotty.hh, vcsn/algos/transpose.hh,
* vcsn/alphabets/setalpha.hh, vcsn/core/mutable_automaton.hh,
* vcsn/core/rat/kratexpset.hh, vcsn/ctx/ctx.hh,
* vcsn/weights/b.hh, vcsn/weights/poly.hh, vcsn/weights/z.hh,
* vcsn/weights/zmin.hh:
(name): Replace by...
(sname, vname): these.