edit-automaton: fix a nasty bug
The dot-parser reads "{2}" as "{2}\e", and uses it as weight.
This is wrong for the initial and final transition, where the
missing label is actually a special-letter.  This is not checked
by mutable_automaton, so we were building an invalid automaton,
and vcsn-cat was not idempotent as automata with initial/final
weight end up showing the "\e".

* TODO.txt: We need a "check" algorithm.
* vcsn/algos/edit-automaton.hh (add_entry): Pay attention to the
special case of initial/final weights.
* tests/unit/vcsn-cat.chk: New.
* tests/unit/local.mk: Use it.