dyn: let context be shared_ptr too
The choice to make dyn::context a genuine structure (but abstract) is
unpleasant, and we have to handle it by pointer, which introduces
nasty memory management issues.  And actually, Valgrind was quite
furious on vcsn-lift for instance (the lifted automaton used the
context of the origin automaton, which was deleted).

Therefore apply the same pattern here: the abstract root is
dyn::abstract_context, and dyn::context is a shared_ptr to it.

* vcsn/ctx/ctx.hh (context): Rename as...
(abstract_context): this.
(context): New.
(make_context): Provide a convenient overload.
* vcsn/ctx/fwd.hh: Adjust.

* lib/vcsn/algos/de_bruijn.cc, lib/vcsn/algos/edit-automaton.cc,
* lib/vcsn/algos/make-context.cc, lib/vcsn/algos/registry.hh,
* lib/vcsn/ctx/ctx.cc, vcsn/algos/aut_to_exp.hh,
* vcsn/algos/edit-automaton.hh, vcsn/algos/make-context.hh,
* vcsn/algos/transpose.hh, vcsn/core/rat/abstract_ratexpset.hh,
* vcsn/dyn/ratexp.hh, vcsn/factory/de_bruijn.hh:
Adjust to dyn::context being now a pointer type.