parse-rat: handle, and return, a dyn::ratexp
This simplifies the code.  It is also simpler and saner to use fewer
types in our interfaces: dyn::ratexp is exactly what we need here,
even if there is so duplication of the context: once as a member of
dyn::ratexp, and another time as a member of the driver.  The latter
might be removable.

* lib/vcsn/rat/driver.hh, lib/vcsn/rat/driver.cc (context_): New member.
Remove the ctor from a ratexpset.
(parse_, parse_string, parse_file): Return a dyn::ratexp.
* lib/vcsn/algos/read.cc, vcsn/core/rat/ratexpset.hxx: Adjust.