context: handle the kind at the labelset level
So far the context was a triple: <Kind, LabelSet, WeightSet>.  This is
troublesome on several regards, the clearest being that the LabelSet
makes no sense for LAU.  Rather, let the contexts be pair: <LabelSet,
WeightSet>, where this time the LabelSet can be a instance of Lau,
Law, or Lal.  These structures, in turn, are parameterized by the
effective set of generators to use (what was named "LabelSet" in the
triple-version of contexts).

This is work in progress.  Scaffolding is still present in the code,
the names (Lau, Lal, Law) are not satisfying (UnitSet, LetterSet and
WordSet seem more appropriate), and some simplification remain to be
done.

* vcsn/ctx/char.hh: Remove.
It used to "factor" the LabelSet (in the former sense:
vcsn::set_alphabet<vcsn::char_letters>) between predefined contexts.
* vcsn/ctx/lau.hh, vcsn/ctx/law.hh, vcsn/ctx/lal.hh: New.
* vcsn/ctx/ctx.hh, vcsn/ctx/fwd.hh (context): Parameterized by
pairs, not triples.
* common.mk, lib/local.mk, vcsn/Makefile.am: Adjust.

* vcsn/ctx/lal_char_b.hh, vcsn/ctx/lal_char_br.hh, vcsn/ctx/lal_char_z.hh,
* vcsn/ctx/lal_char_zmin.hh, vcsn/ctx/lal_char_zr.hh, vcsn/ctx/lal_char_zrr.hh,
* vcsn/ctx/law_char_b.hh, vcsn/ctx/law_char_br.hh, vcsn/ctx/law_char_z.hh,
* vcsn/ctx/law_char_zmin.hh, vcsn/ctx/law_char_zr.hh, vcsn/ctx/law_char_zrr.hh:
Adjust the definitions of the predefined contexts.

* vcsn/ctx/lau_char_br.hh, vcsn/ctx/lau_char_z.hh: rename as...
* vcsn/ctx/lau_br.hh, vcsn/ctx/lau_z.hh: these.
Also change the name of these contexts: the "_char(...)" part is removed.

* vcsn/dyn/ratexp.hh, vcsn/algos/lift.hh, vcsn/algos/xml/structure.hh,
* vcsn/core/mutable_automaton.hh: Adjust.

* vcsn/alphabets/char.hh: Move into Lau what belongs to it.

* tests/unit/aut_lau_char_z.cc. tests/unit/aut_lau_char_z.chk: Rename as...
* tests/unit/aut_lau_z.cc, tests/unit/aut_lau_z.chk: these.
* tests/unit/concat.cc, tests/unit/lift.chk, tests/unit/local.mk,
* tests/unit/transpose.cc:
Adjust.

* NEWS.txt: Document.