lau: remove routines handling words
There is still a lot of code using "word_t" as an "abstraction" of
labels.  LAU is severely touched by this, and trying to clean it
reveals that polynomialset makes too strong assumptions on the parsing
of the labels.  This is clearly something that should be delegated to
the LabelSets.

* vcsn/alphabets/setalpha.hh, vcsn/ctx/letterset.hh,
* vcsn/ctx/unitset.hh, vcsn/ctx/wordset.hh (conv): New.
* vcsn/weights/polynomialset.hh (conv): Use them.
* vcsn/empty.hh (operator<): New, to please polynomialset's map.
* vcsn/ctx/unitset.hh (word_t): Define as empty_t.
Should be removed eventually.