determinize: use hash tables of set of states
Some improvement.  On my machine,

  time vcsn-de-bruijn 17 | vcsn-determinize -f- -Onull

goes from about 18s to about 14s.

A previous experiment stupidly tried to use unordered_set too, and
computed a hash value for it by iteration.  This, of course, results
in a non-constant hash key.  Some commutative operation to combine the
hash should be tried.

* vcsn/algos/determinize.hh: Use unordered map.