make-context: turn assertion into throw
Victor recently uncovered the fact that some vcsn-cat tests were not
properly run.  Once they are run, they abort, instead of reporting the
errors in a more graceful way.

Turn these asserts in if+throw.  While at it, instead of performing a
(dynamic) if on the context type, use enable_if to perform it at
compile time.

* vcsn/algos/make-context.hh (details::make_context): New, fixed,
overloaded splitting of...
(make_context): this.
Bounce to them.