dyn: make_automaton
There are several places where we run make_shared to create a
dyn::automaton.  Fix this by introducing a make_automaton routine, and
use it.  Prepare the future by passing the appropriate context, as, in
the near future, dyn::automaton, like dyn::ratexp, will keep its
context, instead of relying on the fact that mutable_automaton has an
entryset which has a context.

* vcsn/dyn/automaton.hh (make_automaton): New.

* vcsn/algos/accessible.hh, vcsn/algos/aut_to_exp.hh,
* vcsn/algos/complete.hh, vcsn/algos/determinize.hh,
* vcsn/algos/lift.hh, vcsn/algos/product.hh,
* vcsn/algos/standard_of.hh, vcsn/algos/transpose.hh,
* vcsn/factory/de_bruijn.hh, vcsn/factory/ladybird.hh:
Use it.