char_letters: rename identity as empty_word
Currently "identity" is hard-coded in char_letters as the empty word,
which means that it is not a valid label_t when the label_t is simply
letters.

Polynomials seem to really require to work on _words_, so we cannot
simply decide that "identity" would denote a label_t.  So, currently,
the simplest is to rename identity as empty_word.  Nevertheless, we
_must_ go back to deciding what polynomials should really be, and
probably redesign them.

* vcsn/alphabets/char.hh, vcsn/ctx/unitset.hh (identity, is_identity):
Rename as...
(empty_word, is_empty_word): these.
* vcsn/algos/edit-automaton.hh, vcsn/algos/eps-removal.hh,
* vcsn/algos/is-eps-acyclic.hh, vcsn/ctx/genset-labelset.hh,
* vcsn/weights/polynomialset.hh, tests/unit/eps-removal.cc:
Adjust.