internal: introduce entryset to phase out the use of polynomialset
polynomialset is parameterized by a Context, however it does not use
label_t as keys, but word_t.  polynomialset was used for I/O: it
formats nicely together transitions between two states, and is a
convenient abstraction when read such a set of transitions (think of a
GraphViz input for instance).

However, especially in the case of input, using word_t instead of
label_t is a nuisance, as it introduces a useless generalization that
must be controled later.  It is far easier to _always_ read and handle
a label_t.

Yet, polynomialset seems useful, and I don't feel comfortable with
removing it just yet.  So, as a possibly intermediate step, entryset
is almost exactly polynomialset but using label_t keys instead of
word_t.  Some functions cannot be kept, typically all those that are
related to the fact that polynomialset is a WeightSet: unit, star,
etc.  However, to factor common code, it would make sense to have
polynomialset<C1> derive from entryset<C2>, when C2 would be the LAW
context corresponding to C1.

* vcsn/weights/entryset.hh: New.
* vcsn/Makefile.am: Adjust.
* vcsn/algos/dot.hh (format_transition): Use entryset instead of
polynomialset.
* vcsn/ctx/instantiate.hh: No longer instantiate polynomialset.
* vcsn/ctx/nullableset.hh (format, print): New.
We used to rely on their version in char_letters, which works on
word_t.  This is not sane, I prefer to "duplicate" them here.