polynomialset: do not use LabelSet::conv
Currently LabelSet::conv returns a word_t, which sounds really wrong.
This was certainly tailored for polynomialset.  Bypass LabelSet::conv,
and rather use set_alpha::conv, which is able to read words.

A following patch will change the semantics of LabelSet::conv to
return a label_t, not a word_t.

* vcsn/weights/polynomialset.hh: here.