universal: use a functor to avoid stupid bugs due to temporary variables
Recently (see d798139ed68fd5e6cf72bdaf0bb60afef84081f7), universal was
buggy because two different automata coexisted in universal(): the
"raw" automaton, and the one which is complete and deterministic.
At some point, one was used instead of the other.

Instead of having both coexist in the same routine, make this a
functor with two routines: one, private, which _requires_ a complete
deterministic automaton, and another, public, which takes in charge
the needed preconditions.  The core routine then deals with only one
automaton.

The memory management is also simplified.

* vcsn/algos/universal.hh: Do that.