transpose: fix issues with const automata
Basically the issue to solve is to have the following piece of
code compile properly:

  #include <type_traits>

  struct automaton
  {
    bool get_final() const { return final; }
    void set_final()       { final = true; }
    bool final = false;
  };

  template <typename Aut>
  struct transposed_by_hand
  {
    Aut& aut;
    auto get_final() const -> bool
    {
      return aut.get_final();
    }

    auto set_final() -> void
    {
      aut.set_final();
    }
  };

  template <typename Aut>
  struct transposed_with_decltype
  {
    using nocv = typename std::remove_cv<Aut>::type;
    Aut& aut;

    auto get_final() const -> decltype(aut.get_final())
    {
      return aut.get_final();
    }

    auto set_final() -> decltype(const_cast<nocv&>(aut).set_final())
    {
      aut.set_final();
    }
  };

  int main()
  {
    const automaton aut;
    transposed_by_hand<const automaton> trh = { aut };
    transposed_with_decltype<const automaton> trd = { aut };
    trh.get_final();
    //  trh.set_final();

    trd.get_final();
    //  trd.set_final();
  }

where the problem is to provide a systematic means to define
wrapping function (around get_final and set_final in this
small case example) in the wrapper.  The wrapper by hand works
well, but the decltype in the second wrapper fails for set_final
if written this way:

    auto set_final() -> decltype(aut.set_final())
    {
      aut.set_final();
    }

because *in the decltype* clause, aut.set_final() is invalid,
since aut is const.

* vcsn/core/mutable_automaton.hh (self_nocv_t): New type: this
automaton type, but "writable".
* vcsn/algos/transpose.hh: Likewise.
Use const_cast tricks within decltype to avoid issues with forwarding
non-const member functions applied to const objects (see comments
in the code).
* vcsn/algos/determinize.hh: Use self_nocv_t to build a modifiable
automaton from const automata.
* vcsn/algos/accessible.hh, vcsn/algos/universal.hh: Remove ugly
casts.
* tests/unit/transpose.cc: Enable a disabled test.