fsm: start improving the FSM output
There are several issues with the current FSM output: it does not
support labels, weights, handles improperly the special symbol (used
for initial and final transitions) and the difference between acceptor
and transducer.  It is not clear that OpenFST makes a real difference
between transducers and acceptors though.

* tests/checker (check): Support the "ignore" option.
* tests/tafkit/fsm.chk: New.
* tests/tafkit/local.mk: Adjust.
* vcsn/algos/fsm.hh: Be sure to map the special symbol and the one
symbol to 0.
Remove spurious EOL at the end.