labelset: do not rely on setalpha's print and format functions
Labels should not be considered as just letters or words, as
char_letters does.  So instead of just getting print and format
functions from there, define specific ones for letterset and wordset
(nullableset and oneset already have their owns).

* vcsn/alphabets/char.hh: Comment changes.
* vcsn/ctx/genset-labelset.hh (print, format): Do not forward them
from set_alphabet to letterset and wordset.
* vcsn/ctx/letterset.hh, vcsn/ctx/wordset.hh (print, format): New,
just for the right types.
* vcsn/algos/shortest.hh: Adjust: when converting a _word_, use
the set_alphabet, not the LabelSet.