tupleset: conv
Unfortunately this does not work on GCC, which is buggy in its
handling of initializer lists.  The order of evaluation is guaranteed
(by the standard) to be left-to-right, which is vital here since we
have side-effects on the streams.  Luckily, for some reason, output
(print) is properly expanded on both OS X and GNU/Linux, however input
(conv) is not.  As a result, the members of the tuple are not read in
the correct order (and the separators are wrong too, of course: GCC
expects ","s first, and ends with "(").

Clang++ is happily running the test suite.

http://gcc.gnu.org/bugzilla/show_bug.cgi?id=51253

* vcsn/labelset/tupleset.hh (conv, conv_): New.
* tests/unit/label.cc: Check it.