proper: incrementally erase undesirable states
Remove "undesirable" states at each round, instead of waiting for the
end.  So we no longer need a vector to store states to delete.  This
is valid only because we do not add new states, and because
mutable_automaton supports such a pattern.

Also, there is no work to perform at all on states that do not having
incoming spontaneous transitions.

This provides a significant boost.  For instance on
Thompson((\e+a){1000}), we go from 58s to 6s.

* vcsn/algos/proper.hh: here.