proper: extract is-valid in its own file
The proper.hh file is quite complex, with several different kinds of
dispatch (one on the kind, the other on the star status).  It seems
easier to understand and maintain to completely separate the different
components of the "properer" functor.

One goal is also to split the main function of properer
(in_situ_remover) into several functions sharing state via the
properer object.  In that case, the members used to exchange data
between in_situ_remover and its auxiliary functions have nothing to do
with is_valid etc.  So, to avoid that in_situ_remover clutters
is_valid and other routines, move them outside.

* vcsn/algos/fwd.hh: New.
* vcsn/algos/proper.hh: Extract the is_valid parts to...
* vcsn/algos/is-valid.hh: this new file.
Use std::enable_if to dispatch on the star status, instead of
via class template specialization.
* vcsn/Makefile.am: Adjust.