ratexp: transpose: let its visitors work with a ratexpset, not context
The contexts have been overused, to simplify too much.  If for
instance we were to have different types of ratexpsets, what matters
when manipulating a ratexp is not its context, but really its
ratexpset (from which a context can be extracted anyway).

Acknowledge this by first having the ratexp traversals be
parameterized by its ratexpset.

* vcsn/core/rat/transpose.hh, vcsn/core/rat/ratexpset.hxx: Be parameterized
by RatExpSet.
* vcsn/algos/transpose.hh, vcsn/ctx/instantiate.hh: Adjust.