dyn: let ratexps aggregate a ratexpset instead of a ctx
See previous commits: it is far more logical that a dyn::ratexp hold
both a typeful rational expression (i.e., _not_ a vcsn::rat::exp_t,
which is not parameterized by the context), and a typeful ratexpset
(instead of a context).  Anyway, of course, before, dyn::ratexp always
required to build a ratexpset from a context to be used.

As significant consequence, note that now dyn::ratexp have a vname()
corresponding to their ratexpset, not to the context of their
ratexpset.  This changes the instantiation and registration of
algorithms on ratexps.  This, in turn, also highlights that we
neglected some "details" such as the fact that "standard" should
actually be parameterized both by the type of the input ratexp, _and_
the output automaton.

* vcsn/dyn/ratexp.hh (abstract_ratexp::ctx_, abstract_ratexp::ratexp_):
Remove.
(abstract_ratexp::vname, abstract_ratexp::as): New.
(concrete_abstract_ratexp): New.
Aggregates typeful ratexp and ratexpset.
(make_ratexp): Adjust.
* vcsn/dyn/ratexpset.hh (ratexpset<Context>::make_ratexp): New.

* lib/vcsn/rat/driver.cc (parse_): Use the previous make_ratexp
to return the dyn::ratexp.
* vcsn/core/rat/ratexpset.hxx (conv): Adjust the call to the
rat parsing driver.

* vcsn/algos/aut_to_exp.hh, vcsn/algos/info.hh, vcsn/algos/lift.hh,
* vcsn/algos/print.hh, vcsn/algos/standard.hh, vcsn/algos/thompson.hh,
* vcsn/algos/transpose.hh, vcsn/algos/constant-term.hh:
Adjust to the new nature of dyn::ratexp: be parameterized by a
RatExpSet instead of a context, and more generally, avoid going
via a context when not needed.

* lib/vcsn/algos/standard.cc, lib/vcsn/algos/thompson.cc,
* lib/vcsn/algos/constant-term.cc:
Now dispatch on the vname of the dyn::ratexp itself, instead of
that of its context (it no longer comes with one anyway).

* vcsn/ctx/instantiate.hh: Adjust: register under the name of
the ratexpset the previous algorithms.