proper: speed up: use a heuristic to decide the order of state processing
After discussing with Jacques Sakarovitch, it became clear that the
current approach is way too naive.

It eliminates states in their "implementation" order (that of their
state_t index), which is totally unrelated to the topology of the
automaton itself.  This can result in introducing many useless
spontaneous transitions that are eliminated later.  Think for instance
of the applying proper to thompson(a?{n}): it is much more efficient
to work "from final to initial states", than the converse (which is
what the "implementation order" actually does.  For n=2000,
the "implementation order" takes 102s on my machine.

A better idea (based on suggestions from Jacques), similar to what is
done in aut-to-exp, consists in eliminating first states with fewer
outgoing spontaneous transitions, then states with fewer outgoing
transitions, and finally states with fewer incoming spontaneous
transitions.

Obviously, since these values change when a state is processed (it
changes the in/out transitions of its neighbors) the order must be
updated.  To this end we will use a heap.

However this turns out to be very costly if implemented naively:
updating the heap for each neighbor of the processed node involves
many comparisons with other states.  In the end, there are too many
iterations on the in/out transitions of the states, to compute this
order.  So we introduce a "cache" for these values (number of in/out
transition per node), named "state_profile", which we update after
each processing of a state, right before updating the heap.

On my machine, this cuts down proper(thompson(a?{2000})) from 102s to
15s.

* vcsn/algos/proper.hh (debug_level): New.
(state_profile, update_profile_): New.
(build_heap_, update_heap_, show_heap_): New.
(in_situ_remover_(state_t)): Simplify signature.
(in_situ_remover_()): Use and maintain the heap.
* tests/tafkit/proper.chk: Fix expectations: we get a different, but
equivalent result (actually, a nicer one).