news: beware of buggy Makefile dependencies
On the buildfarm several tests are failing because they were not
upgraded properly.  For instance:

  node8 $ ls -ltr vcsn/core/mutable_automaton.hh build/gcc48/tests/demo/*
  -rw-r--r-- 1 forge forge 1975672 Sep 24 16:55 build/gcc48/tests/demo/prod-eval.o
  -rw-r--r-- 1 forge forge   20030 Sep 26 18:09 vcsn/core/mutable_automaton.hh
  -rwxr-xr-x 1 forge forge    9084 Sep 27 09:59 build/gcc48/tests/demo/prod-eval

i.e., although get_transition was changed in mutable_automaton.hh (and
in all the libraries), the prod-eval.o object was not updated.  This
is because its .Po is about "demo/prod-eval.o" (its former name)
instead of "tests/demo/prod-eval.o".  In otherword, the Makefile
properly loads the Po file of tests/demo/prod-eval.o, but the latter
contains dependencies about another file, so the file is not updated
again.

* NEWS.txt: Add a warning.