labelset: define "value_t" for consistency, in addition to label_t
In some cases, for instance polynomials of ratexps, the labelset does
not define label_t, or defines it to something different than its
members.  For instance in the case of ratexpset, label_t is the labels
of the ratexps themselves, whereas we need a reference to members of
the ratexpsets, i.e., the ratexp themselves.

To avoid such issues, implementation of element/elementset should have
elementset provide a "value_t" type which denotes the type of its
elements.  For a start, let's convert the labelsets, and their uses.

Eventually, these labelsets should no longer define label_t, as it is
misleading.

Note that for similar reasons, it is important that "one()" be the
sole name to denote the neutral element: in the case of polynomials of
ratexp, the "one" polynomial uses the "one" ratexp, whereas the code
is used to ask for the "empty_word()" of the labelset.

* vcsn/labelset/letterset.hh, vcsn/labelset/oneset.hh,
* vcsn/labelset/wordset.hh (value_t): New.
* vcsn/weights/polynomialset.hh, vcsn/weights/entryset.hh: Use it.