polynomialset: use static local variables instead of members for one and zero
This change somewhat saves space (one and zero are allocated only once
per polynomialset type, not per instance), saves time (costs once, not
at each polynomialset instantiation), but these are minor (probably
even insignificant) details.

What does matter though is that now zero_ and one_ are instantiated
only when needed.  In particular, this allows to use polynomialset
with labelsets that do not have a one() label (e.g., LAL).

In other words, entryset is no longer needed, as polynomialset is a
superset of its features, and also applies to the cases where entryset
was believed to be needed.

* vcsn/weights/polynomialset.hh (zero_, one_): Move their definitions to...
(zero, one): here.
Adjust the constructor.