dyn: simplify the instantiation of the bridges
There is really no reason to force the bridges to be overloaded (e.g.,
using "standard" for two routines, one with automaton as input, and
the other with "ratexp" as input), as these functions are never seen
by the user.

However, it does make the instantiatiation and use by the system more
complex (e.g., the REGISTER macros cannot be used).

Relax these constraints, and use the same name for the bridge, and for
its associated typedef (standard -> standard_t, standard_exp ->
standard_exp_t).

* vcsn/algos/edit-automaton.hh, vcsn/factory/ladybird.hh: Comment change.
* vcsn/algos/print.hh, vcsn/algos/standard.hh: Use the same names for
bridges and types.
* vcsn/ctx/lan_char_b.hh, vcsn/ctx/instantiate.hh: Simplify.