build: control the size of the library symbol table
Instead of explicitly hiding the symbols that must not be exported, do
the converse: require that by default symbols are _not_ exported, but
in turn explicitly flag what symbols must be exported.

The vcsn/misc/export.hh file provide a symbol, LIBVCSN_API, which must
be explicitly used to flag symbols to export, _and_ used to import
them in callers.

Of course all the dyn/algos.hh signature must be exported, which is
easy to write thanks to the namespace attributes syntax from GCC.
Likewise for algorithms in vcsn/misc that are actually implemented in
*.cc files.

What was not really expected is that explicit class template
instantiations must also be flagged with LIBVCSN_API (see for instance
its uses in vcsn/ctx/instantiate.hh).

On my machine, using 4.8, I get a significant reduction of the library
sizes (roughly 20%):

  920K -> 732K liblal_char_b.0.dylib
  1,1M -> 788K liblal_char_br.0.dylib
  840K -> 660K liblal_char_q.0.dylib
  828K -> 656K liblal_char_r.0.dylib
  840K -> 660K liblal_char_z.0.dylib
  848K -> 664K liblal_char_zmin.0.dylib
  1,1M -> 800K liblal_char_zr.0.dylib
  1,2M -> 912K liblal_char_zrr.0.dylib
  864K -> 680K liblan_char_b.0.dylib
  844K -> 660K liblan_char_r.0.dylib
  832K -> 648K liblan_char_z.0.dylib
  1,1M -> 812K liblan_char_zr.0.dylib
  816K -> 628K liblao_br.0.dylib
  572K -> 464K liblao_z.0.dylib
  768K -> 604K liblaw_char_b.0.dylib
  964K -> 740K liblaw_char_br.0.dylib
  796K -> 628K liblaw_char_q.0.dylib
  780K -> 620K liblaw_char_r.0.dylib
  768K -> 608K liblaw_char_z.0.dylib
  780K -> 616K liblaw_char_zmin.0.dylib
  968K -> 748K liblaw_char_zr.0.dylib
  1,1M -> 848K liblaw_char_zrr.0.dylib
  764K -> 720K libvcsn.0.dylib

   20M -> 16M  TOTAL

There is no compilation speed-up (or I failed to measure it).  However
we _might_ benefit from a small speed-up when launching TAF-Kit (as the
dynamic loader as a simpler job to do):

Before:
  8.53s (0.08s+8.45s): ladybird 21 | determinize -O null
  0.17s (0.08s+0.09s): de-bruijn 20 | evaluate -O null -f - aaaaaaaaaaaaaaaaaaaaa
 12.84s (1.03s+11.81s): thompson -C "lan_char(a)_b" -Ee "a?{2000}" | proper -O null
 23.30s (10.37s+12.93s): standard -Ee "(a+b+c+d)?{100}" | aut-to-exp -O null
  7.27s (0.07s+7.20s): standard -C "lal_char(ab)_z" -Ee "(a+b)*b(<2>a+<2>b)*" | power -O null -f- 10
  0.07s (0.07s): standard -E -e "(a?){70}" -o a70.gv
 13.63s (13.63s): product -O null -f a70.gv a70.gv

After:
  7.94s (0.05s+7.89s): ladybird 21 | determinize -O null
  0.10s (0.05s+0.05s): de-bruijn 20 | evaluate -O null -f - aaaaaaaaaaaaaaaaaaaaa
 12.04s (0.93s+11.11s): thompson -C "lan_char(a)_b" -Ee "a?{2000}" | proper -O null
 22.69s (10.16s+12.53s): standard -Ee "(a+b+c+d)?{100}" | aut-to-exp -O null
  6.72s (0.05s+6.67s): standard -C "lal_char(ab)_z" -Ee "(a+b)*b(<2>a+<2>b)*" | power -O null -f- 10
  0.05s (0.05s): standard -E -e "(a?){70}" -o a70.gv
 13.01s (13.01s): product -O null -f a70.gv a70.gv

Granted, the differences are small and might be insignificant.

* configure.ac: Check if the compiler supports -fvisibility=hidden.
* lib/local.mk, lib/vcsn/local.mk: Use it to compile libvcsn and
the context libraries.

* lib/ctx/libctx.cc: Enable the exportation of the symbols.

* vcsn/misc/export.hh: New.
* vcsn/local.mk: Adjust.

* lib/vcsn/dot/driver.hh, lib/vcsn/rat/driver.hh,
* lib/vcsn/rat/read.hh, vcsn/core/mutable_automaton.hh,
* vcsn/ctx/instantiate.hh, vcsn/dyn/algos.hh, vcsn/dyn/automaton.hh,
* vcsn/dyn/context.hh, vcsn/dyn/fwd.hh, vcsn/dyn/ratexp.hh,
* vcsn/dyn/weight.hh, vcsn/misc/escape.hh, vcsn/misc/stream.hh:
Flag the symbols that need to be exported.