dyn: use better names
Currently we have abstract_foo, which is the root of all foos, and
concrete_abstract_foo, which wraps a foo.  This is not very clear, to
remain polite.

It is the tradition in C++ to use "base" for superclasses, so let's
first rename abstract_foo as "foo_base" (like ios_base).  Note that we
keep the "foo_" part as a prefix, instead of a suffix.

Then, concrete_abstract_foo derives from abstract_foo (which makes it
somewhat concrete), but foo also derives from abstract_foo.  It's real
nature is to _wrap_ a genuine foo.  So let's name it "foo_wrapper".

* doc/vcsn2.xmi, lib/vcsn/algos/edit-automaton.cc,
* vcsn/algos/edit-automaton.hh, vcsn/algos/transpose.hh,
* vcsn/core/mutable_automaton.hh, vcsn/core/rat/ratexpset.hh,
* vcsn/core/rat/ratexpset.hxx, vcsn/dyn/automaton.hh, vcsn/dyn/fwd.hh,
* vcsn/dyn/polynomial.hh, vcsn/dyn/polynomialset.hh,
* vcsn/dyn/ratexp.hh, vcsn/dyn/ratexpset.hh, vcsn/dyn/ratexpset.hxx,
* vcsn/dyn/weight.hh, vcsn/dyn/weightset.hh, vcsn/weights/b.hh,
* vcsn/weights/f2.hh, vcsn/weights/polynomialset.hh,
* vcsn/weights/q.hh, vcsn/weights/r.hh, vcsn/weights/z.hh,
* vcsn/weights/zmin.hh:
s/concrete_abstract_((automaton|polynomial|ratexp|weight)(set)?)/$1_wrapper/g.
         s/abstract_((automaton|polynomial|ratexp|weight)(set)?)/$1_base/g.