dyn: let dyn::context _wrap_ a context
Contrary to all the other dyn:: implementation, dyn::context was
actually a (pointer to a) genuine static context.  This requires that
static contexts derive from dyn::detail::context_base, which is the
last case of a class of the static API deriving from the a class of
the dyn API.

Introduce context_wrapper instead, and use the same scheme as
elsewhere.

* vcsn/dyn/context.hh (context_wrapper): New.
(abstract_context): Rename as...
(context_base): this.
(as): Adjust.
* vcsn/ctx/context.hh (context): No longer derive from context_base.
Adjust.

* lib/vcsn/algos/make-context.cc, lib/vcsn/ctx/ctx.cc,
* tests/demo/prod-eval.cc, vcsn/algos/edit-automaton.hh,
* vcsn/algos/make-context.hh, vcsn/dyn/fwd.hh,
* vcsn/factory/de-bruijn.hh, vcsn/factory/divkbaseb.hh,
* vcsn/factory/double-ring.hh, vcsn/factory/ladybird.hh,
* vcsn/factory/random.hh, vcsn/factory/u.hh:
Adjust.