dyn: simplify make_automaton
make_automaton was reworked to take a context, although it it useless,
thinking about the future: the plan was to have dyn::automaton wrap
both a context, and a static automaton.  Currently this is useless
because the static automaton already wraps a context.

However, it becomes clearer to me that the proper design is to have
dyn::automaton wrap both a static automaton, and an automatonset.
Yes, automatonsets do not exist, however, I start to believe that it
was an error not to apply the "element/elementset" design to automata
too.

* vcsn/dyn/automaton.hh (make_automaton): No longer take a context.

* vcsn/algos/accessible.hh, vcsn/algos/aut-to-exp.hh,
* vcsn/algos/complement.hh, vcsn/algos/complete.hh,
* vcsn/algos/concatenate.hh, vcsn/algos/copy.hh,
* vcsn/algos/derivation.hh, vcsn/algos/determinize.hh,
* vcsn/algos/left-mult.hh, vcsn/algos/lift.hh, vcsn/algos/product.hh,
* vcsn/algos/proper.hh, vcsn/algos/standard.hh, vcsn/algos/star.hh,
* vcsn/algos/sum.hh, vcsn/algos/thompson.hh, vcsn/algos/transpose.hh,
* vcsn/algos/union.hh, vcsn/algos/universal.hh,
* vcsn/factory/de-bruijn.hh, vcsn/factory/divkbaseb.hh,
* vcsn/factory/double-ring.hh, vcsn/factory/ladybird.hh,
* vcsn/factory/random.hh, vcsn/factory/u.hh:
Adjust.